requirements.txt
discord.py
flask
pytz
motor
python-dotenv